Created separate projects for services that should only be used internally by other Neo.UI.Core projects
