+++ Core/Neo.UI.Core.Internal.Services.Implementations/Neo.UI.Core.Internal.Services.Implementations.csproj
@@ -12,8 +12,8 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Neo.UI.Core.Data\Neo.UI.Core.Data.csproj" />
-    <ProjectReference Include="..\Neo.UI.Core.Helpers\Neo.UI.Core.Helpers.csproj" />
     <ProjectReference Include="..\Neo.UI.Core.Internal.Services.Interfaces\Neo.UI.Core.Internal.Services.Interfaces.csproj" />
+    <ProjectReference Include="..\Neo.UI.Core.Messaging\Neo.UI.Core.Messaging.csproj" />
     <ProjectReference Include="..\Neo.UI.Core.Services.Interfaces\Neo.UI.Core.Services.Interfaces.csproj" />
   </ItemGroup>
