+++ Core/Neo.UI.Core.Internal.Services.Interfaces/ICertificateQueryService.cs
