+++ Core/Neo.UI.Core.Internal.Services.Interfaces/Neo.UI.Core.Internal.Services.Interfaces.csproj
@@ -12,9 +12,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Neo.UI.Core.Data\Neo.UI.Core.Data.csproj" />
-    <ProjectReference Include="..\Neo.UI.Core.Helpers\Neo.UI.Core.Helpers.csproj" />
-    <ProjectReference Include="..\Neo.UI.Core.Internal.Services.Interfaces\Neo.UI.Core.Internal.Services.Interfaces.csproj" />
-    <ProjectReference Include="..\Neo.UI.Core.Services.Interfaces\Neo.UI.Core.Services.Interfaces.csproj" />
+    <ProjectReference Include="..\Neo.UI.Core.Theming\Neo.UI.Core.Theming.csproj" />
   </ItemGroup>
 
 </Project>
