+++ Core/Neo.UI.Core.Wallet/Neo.UI.Core.Wallet.csproj
@@ -15,6 +15,7 @@
     <ProjectReference Include="..\Neo.UI.Core.Globalization\Neo.UI.Core.Globalization.csproj" />
     <ProjectReference Include="..\Neo.UI.Core.Data\Neo.UI.Core.Data.csproj" />
     <ProjectReference Include="..\Neo.UI.Core.Helpers\Neo.UI.Core.Helpers.csproj" />
+    <ProjectReference Include="..\Neo.UI.Core.Internal.Services.Implementations\Neo.UI.Core.Internal.Services.Implementations.csproj" />
     <ProjectReference Include="..\Neo.UI.Core.Messaging\Neo.UI.Core.Messaging.csproj" />
     <ProjectReference Include="..\Neo.UI.Core.Services.Implementations\Neo.UI.Core.Services.Implementations.csproj" />
     <ProjectReference Include="..\Neo.UI.Core.Transactions\Neo.UI.Core.Transactions.csproj" />
