+++ Core/Neo.UI.Core.Wallet/WalletRegistrationModule.cs
@@ -4,6 +4,8 @@ using Neo.UI.Core.Messaging;
 using Neo.UI.Core.Services.Implementations;
 using Neo.UI.Core.Transactions;
 
+using InternalServicesRegistrationModule = Neo.UI.Core.Internal.Services.Implementations.ServicesRegistrationModule;
+
 namespace Neo.UI.Core.Wallet
 {
     public class WalletRegistrationModule : Module
@@ -16,6 +18,7 @@ namespace Neo.UI.Core.Wallet
                 .SingleInstance();
 
             // Register modules
+            builder.RegisterModule<InternalServicesRegistrationModule>();
             builder.RegisterModule<MessagingRegistrationModule>();
             builder.RegisterModule<ServicesRegistrationModule>();
             builder.RegisterModule<TransactionsRegistrationModule>();
